adjusting the plotting function and trying to make it more reproducible
edited the tri_plot_bias_dierction() function as it was cutting off some of the plots. also tried to make it more reproducible with a variety of number of studies by dynamically scaling the vertical buffer based on the number of studies and subgroups. not sure how much difference it makes but i have added a vignette which experiments with example datasets. i will change the format of everything as theres now 2 plotting functions but let me know what you think.